adding custom response for all

+++ src/main/java/com/thinkdestructive/Restload/service/impl/CloudVendorServiceImpl.java
@@ -1,6 +1,8 @@
 package com.thinkdestructive.Restload.service.impl;
 
+import com.thinkdestructive.Restload.exception.CloudVendorException;
 import com.thinkdestructive.Restload.exception.CloudVendorNotFoundException;
+import com.thinkdestructive.Restload.exception.NoCloudVendorExistException;
 import com.thinkdestructive.Restload.model.CloudVendor;
 import com.thinkdestructive.Restload.repository.CloudVendorRepository;
 import com.thinkdestructive.Restload.service.CloudVendorService;
@@ -22,15 +24,21 @@ public class CloudVendorServiceImpl implements CloudVendorService {
 
     @Override
     public String updateCloudVendor(CloudVendor cloudVendor) {
-        cloudVendorRepository.save(cloudVendor);
-        return "Success";
+        if(cloudVendorRepository.findById(cloudVendor.getVendorId()).isPresent()) {
+            cloudVendorRepository.save(cloudVendor);
+            return "Success";
+        }
+        throw new CloudVendorNotFoundException("No such cloudVendor exist with given vendor Id = "+cloudVendor.getVendorId());
     }
 
     @Override
     public String deleteCloudVendor(String cloudVendorId) {
-        cloudVendorRepository.deleteById(cloudVendorId);
-        return "Success";
-    }
+        if(cloudVendorRepository.findById(cloudVendorId).isPresent()){
+            cloudVendorRepository.deleteById(cloudVendorId);
+            return "Success";
+        }
+        throw new CloudVendorNotFoundException("No such cloudVendor exist with vendor Id = "+cloudVendorId);
+}
 
     @Override
     public CloudVendor getCloudVendor(String cloudVendorId) {
@@ -40,15 +48,22 @@ public class CloudVendorServiceImpl implements CloudVendorService {
     }
     @Override
     public List<CloudVendor> getByVendorName(String vendorName) {
+        if (cloudVendorRepository.findByVendorName(vendorName).isEmpty())
+            throw new CloudVendorNotFoundException("CloudVendor Name doesn't Match with any existing vendorName.");
         return cloudVendorRepository.findByVendorName(vendorName);
-//        if (vendors.isEmpty()) {
-//            throw new CloudVendorNotFoundException("Requested CloudVendor doesn't exist.");
-//        }
-//        return vendors;
     }
 
     @Override
     public List<CloudVendor> getAllCloudVendor() {
+        System.out.println(cloudVendorRepository.findAll());
+        if(cloudVendorRepository.findAll().isEmpty()){
+            throw new NoCloudVendorExistException("No Vendor Exist in the DB.");
+        }
         return cloudVendorRepository.findAll();
+    }
+
+    public String deleteAllCloudVendor(){
+        cloudVendorRepository.deleteAllInBatch();
+        return "Success";
     }
 }
